% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme_en.R
\docType{data}
\name{my_theme_en}
\alias{my_theme_en}
\title{Custom theme for English tables}
\format{
An object of class \code{list} of length 14.
}
\usage{
my_theme_en
}
\value{
A list of custom gtsummary theme options.
}
\description{
This function creates a custom theme for gtsummary tables in English.
}
\examples{
gtsummary::set_gtsummary_theme(my_theme_en)

}
\seealso{
my_theme_spa
}
\keyword{custom-theme}
\keyword{english}
\keyword{gtsummary}
\keyword{tables}
