% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme_en.R
\name{my_theme_en}
\alias{my_theme_en}
\title{Custom theme for English tables}
\usage{
my_theme_en()
}
\value{
A list of custom gtsummary theme options.
}
\description{
This function creates a custom theme for gtsummary tables in English.
}
\examples{
set_gtsummary_theme(my_theme_en)

}
